+++ src/richie/apps/core/helpers.py
@@ -128,7 +128,7 @@ def recursive_page_creation(site, pages_info, parent=None):
     for name, kwargs in pages_info.items():
         children = kwargs.pop("children", None)
         try:
-            page = Page.objects.get(reverse_id=name)
+            page = Page.objects.get(reverse_id=name, publisher_is_draft=False)
         except Page.DoesNotExist:
             page = create_i18n_page(
                 site=site, parent=parent, published=True, reverse_id=name, **kwargs